+++ src/main/java/answers/lukasz/Tasks/Task3/Task3_Array.java
@@ -1,4 +1,4 @@
-package answers.lukasz.Tasks;
+package answers.lukasz.Tasks.Task3;
 
 import java.util.Scanner;
