+++ src/main/java/answers/lukasz/Tasks/Task4/Task4_1.java
@@ -1,4 +1,4 @@
-package answers.lukasz.Tasks;
+package answers.lukasz.Tasks.Task4;
 
 import java.util.Scanner;
