+++ src/main/java/answers/lukasz/Tasks/Task8/Main.java
@@ -1,4 +1,4 @@
-package answers.lukasz.Tasks;
+package answers.lukasz.Tasks.Task8;
 import java.util.Scanner;
 public class Main {
     public static void main(String[] args) {
